refactor: move step into separate components rather than single-file
